Added redux and firestore integration with event blocks
