+++ src/components/MonthCalendar.js
@@ -39,7 +39,7 @@ const CalendarItem = (props) => {
         }
     }
 
-    let sortedBlocks = props.blocks;
+    let sortedBlocks = [...props.blocks];
     if (sortedBlocks) {
         sortedBlocks.sort((a,b) => {
             return a.yPos - b.yPos;
