+++ src/redux_functions/actionTypes.js
@@ -5,5 +5,7 @@ export const Types = {
     SIGNUP: 'SIGNUP',
     SIGNUP_ERROR: 'SIGNUP_ERROR',
     ADD_EVENT: 'ADD_EVENT',
-    EDIT_BLOCK: 'EDIT_BLOCK'
+    EDIT_BLOCK: 'EDIT_BLOCK',
+    DELETE_BLOCK: 'DELETE_BLOCK',
+    CLEAR_BLOCKS: 'CLEAR_BLOCKS'
 }
