+++ src/redux_functions/reducers.js
@@ -63,6 +63,7 @@ export function authReducer(state=initialAuthState, action) {
 }
 
 export function blocksReducer(state = initialBlocksState, action) {
+    let newState;
     switch(action.type) {
         case Types.ADD_EVENT:
             return [
@@ -70,13 +71,26 @@ export function blocksReducer(state = initialBlocksState, action) {
                 action.payload.block
             ];
         case Types.EDIT_BLOCK:
-            let newState = [...state];
+            newState = [...state];
             for (let block of newState) {
                 if (block.key === action.payload.block.key) {
-                    block = action.payload.block;
+                    block = {
+                        ...block,
+                        ...action.payload.block
+                    };
                 }
             }
             return newState;
+        case Types.DELETE_BLOCK:
+            newState = [];
+            for (let block of state) {
+                if (block.key !== action.payload.blockID) {
+                    newState.push({...block});
+                }
+            }
+            return newState;
+        case Types.CLEAR_BLOCKS:
+            return [];
         default:
             return state;
     }
